move help-tooltip to shared dir
